Add a return value to Packet::RawSocketSend
Instead of call exit when the sendto function returns a value
less than zero, the users should handle the return value of the
Packet::RawSocketSend method. In case the function sendto returns
a values less than zero, the library prints the associated "perror"
message (but the user can ignore it).