Better management of open raw socket
Instead of opening and closing the socket each time a Send*() function is called,
the Packet class keep track of the already open sockets. If there aren't changes
on the way a raw socket is open, the Send*() functions reuse an already open
socket. If not, close it and open another one.