Better management of open sockets.
The library now has a "global" socket manager that open only one
socket for each interface/protocol (PF_PACKET or PF_INET) that the
user wants to send. Write operation on sockets are thread-safe so there
isn't a problem to share a few sockets descriptor between an arbitrary
number of threads.