Add initial support for TCP Options
Now is possible to craft and parse TCP options with libcrafter. A very easy
interface to add new options was developed, and also a few options were
implemented.

However, the TCP Craft method should be modified to take into account the
TCP Options layers on the top of TCP to automatically fill the Offset field.