Packet decoder code is independent of Layer classes
Now each Layer child has information for construction from raw data on
the virtual function ParseLayerData. The PacketDecoder code is
independent of specific Layer classes. This results on a more flexible
and mantainable code.

The code on PacketDecoder.cpp shouldn't be changed when a new Protocol
is added resulting in a more modular development.