New Bind function
Now is possible to "bind" a layer on the top of another layer. This is
useful when dealing with layer 5 protocols, like DNS or DHCP.

For example, is possible to configure the library on run-time to put a DNS
layer on the top of an UDP packet with src or dst port 53 when is
parsing a packet from the net.