Add Packet::PreCraft() method
When the PreCraft method is called a packet is crafted. Since then, the
crafting of that packet is full responsability of the user. Libcrafter
won't fill any field after that.

This method could be useful to send a lot of packet on a short period of
time (because the Craft method won't be called inside the Send*
methods).